Remove old tests from webserver workflow
The webserver container used to build a frontend with node and needed to
run tests on the JavaScript. Now all of that is part of the app
workflows and this is just for creating the container that runs the
nginx reverse proxy, so we don't need those commented-out jobs anymore.