Update the scalar variable responses
Rewrite the functions involved in the scalar variable diagnostics routes
to stream the responses as a stripped down JSON format. We're now able
to use generators so that we only have to jsonify small dictionaries and
stream each chunk of JSON in the response, which is significantly faster
than parsing the entire response from an in-memory data structure into
JSON and then sending that all at once.