Change vector tests to expect flattened columns
We can more efficiently generate the JSON response for vector variables
if we flatten the vector columns instead of having a nested dictionary
containing the vector components for the key. Arguably, this flatter
structure is better for the frontend, too, since we can just map x and y
to properties on each record without having to implement any kind of
JavaScript path parsing.