Use DataFrame.to_json to generate diag responses
This is much faster and easier than the generator approach we were using
before. Mainly because we handle vectors by just flattening the column
hierarchy of the vector components instead of trying to merge the u and
v columns into a single dictionary for each column.

There's a decent chance that all of this hassle could be avoided by
storing the vector data differently. Or perhaps this is for the best
because it makes it easier to handle the vector columns when we
visualize them because our chart components don't have to worry about
parsing JavaScript object paths like "obs_minus_forecast_adjusted.u".