Add a default 2 week range for the historical data
This is an attempt to keep memory usage under control in production
where we have a long history of data, so we can’t just show you the
history of all the model runs for which we have records. For now, we
have a hard-coded limit of two weeks of data, but eventually we may
parameterize the start and end dates.