Fix test data to use real dates, not strings
This is an example of our test setup becoming out of sync with how our
data pipeline actually writes the data, I think. The tests were passing
comparing strings, but the actual application failed, because our real
data is written out as timestamps. I updated the tests to ensure we're
creating Parquet files with timestamps, and fixed the comparison in the
diag.history function. Now the application appears to work as well as
all the tests passing.