Apply the is_used filters on load
Moving the is_used filter to the Parquet loading is significantly
improving performance for me locally when loading the wind data.
Ideally, I think it would be good to let PyArrow do all of the
filtering, so that may be worth looking into.

I had to update the test data in a number of cases to ensure that the
is_used flag is a bool, not an int.