Explicitly unstack the component index
For some reason my test data in development has its index ordered
(component, nobs) instead of (nobs, component). Not sure if this is
indicative of a bug somewhere in our ETL code, or if I messed something
up generating this data. Regardless, being specific like this will work
no matter what, so I think we should be ok with this for now.