Copy ChartTimeseries for ChartObsCount
Copy-pasta the code for the ChartTimeseries web component to a new
component - ChartObsCount - that displays the number of used
observations over time. I'm struggling to come up with a set of
components that allows us to create more generalized charts without
hard-coding data access etc. This is an extremely stupid solution to
this problem, but it moves the project forward without over-engineering
the problem by inventing a whole charting primitives library from
scratch.