Fit the map to the data
Create a GeoJSON FeatureCollection from the data that we can pass to
D3's fitSize method to zoom the map to the data.

I tried using d3.extent to find the range of the longitude and latitude
in the data from which I created a FeatureCollection with just 2 points,
but the longitude range ended up being nearly [-180, 180], so it never
actually zoomed the map. Seems like maybe D3 is doing something smart to
handle collections that wrap around the globe, or possibly there's some
unaccounted for weirdness in the data.